requirements.txt
absl-py>=2.1.0
adversarial-robustness-toolbox>=1.15.0
albumentations>=1.3.1
annotated-types>=0.7.0
anyio>=4.3.0
autoawq>=0.1.8
bitsandbytes>=0.43.0
certifi>=2024.2.2
charset-normalizer>=3.3.2
click>=8.1.7
cma>=3.2.2
colorama>=0.4.6
contourpy>=1.2.1
cycler>=0.12.1
datasets>=2.18.0
distro>=1.9.0
fonttools>=4.51.0
graphviz>=0.20.3
h11>=0.14.0
httpcore>=1.0.5
httpx>=0.27.0
idna>=3.7
Jinja2>=3.1.4
joblib>=1.4.2
kiwisolver>=1.4.5
lpips>=0.1.4
MarkupSafe>=2.1.5
matplotlib>=3.8.0
nltk>=3.8.1
numpy>=1.26.4
num2words>=0.5.14
opencv-python>=4.8.0
openai>=1.30.1
packaging>=24.0
pandas>=2.1.2
pillow>=10.3.0
pip>=24.0
prettytable>=3.10.0
psutil>=5.9.0
pydantic>=2.7.1
pydantic_core>=2.18.2
pyecharts>=2.0.5
pyparsing>=3.1.2
python-dateutil>=2.9.0.post0
python-dotenv>=1.0.1
pytz>=2024.1
qwen-vl-utils[decord]>=0.0.8
regex>=2024.5.15
requests>=2.32.2
rouge>=1.0.1
rouge_score>=0.1.2
scikit-image>=0.20.0
seaborn>=0.13.2
setuptools>=69.5.1
simplejson>=3.19.2
six>=1.16.0
snapshot-phantomjs>=0.0.3
sniffio>=1.3.1
spacy>=3.7.4
tiktoken>=0.7.0
torch>=2.2.0
torchvision>=0.18.0
tqdm>=4.66.4
# Use the latest transformers from source for TinyLLaVA support
git+https://github.com/huggingface/transformers.git
accelerate>=0.25.0
tabulate
typing_extensions>=4.11.0
tzdata>=2024.1
urllib3>=2.2.1
wcwidth>=0.2.13
wheel>=0.43.0
# CLIP for semantic similarity in adversarial attacks
git+https://github.com/openai/CLIP.git
# PDF extraction dependencies
pymupdf
camelot-py[cv]
ghostscript
tabulate
markitdown
# DeepSeek-VL2 dependencies
git+https://github.com/deepseek-ai/DeepSeek-VL.git
# # TinyLLaVA dependencies
# git+https://github.com/TinyLLaVA/TinyLLaVA_Factory.git
# Optional: Flash Attention 2 for better performance with Qwen2-VL models
# flash-attn>=2.5.0
